Move syncState repo/entity/dao to package
