+++ stream-chat-android-offline/api/stream-chat-android-offline.api
@@ -237,12 +237,6 @@ public final class io/getstream/chat/android/livedata/dao/QueryChannelsDao_Impl 
 	public fun select (Ljava/util/List;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
-public final class io/getstream/chat/android/livedata/dao/SyncStateDao_Impl : io/getstream/chat/android/livedata/dao/SyncStateDao {
-	public fun <init> (Landroidx/room/RoomDatabase;)V
-	public fun insert (Lio/getstream/chat/android/livedata/entity/SyncStateEntity;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun select (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
 public final class io/getstream/chat/android/livedata/extensions/ClientExtensionsKt {
 	public static final fun isPermanent (Lio/getstream/chat/android/client/errors/ChatError;)Z
 }
@@ -256,7 +250,7 @@ public final class io/getstream/chat/android/livedata/repository/database/ChatDa
 	public fun messageDao ()Lio/getstream/chat/android/livedata/repository/domain/message/MessageDao;
 	public fun queryChannelsQDao ()Lio/getstream/chat/android/livedata/dao/QueryChannelsDao;
 	public fun reactionDao ()Lio/getstream/chat/android/livedata/repository/domain/reaction/ReactionDao;
-	public fun syncStateDao ()Lio/getstream/chat/android/livedata/dao/SyncStateDao;
+	public fun syncStateDao ()Lio/getstream/chat/android/livedata/repository/domain/syncState/SyncStateDao;
 	public fun userDao ()Lio/getstream/chat/android/livedata/repository/domain/user/UserDao;
 }
 
@@ -288,6 +282,12 @@ public final class io/getstream/chat/android/livedata/repository/domain/reaction
 	public fun selectSyncNeeded (Lio/getstream/chat/android/client/utils/SyncStatus;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun selectUserReactionsToMessage (Ljava/lang/String;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun setDeleteAt (Ljava/lang/String;Ljava/lang/String;Ljava/util/Date;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
+public final class io/getstream/chat/android/livedata/repository/domain/syncState/SyncStateDao_Impl : io/getstream/chat/android/livedata/repository/domain/syncState/SyncStateDao {
+	public fun <init> (Landroidx/room/RoomDatabase;)V
+	public fun insert (Lio/getstream/chat/android/livedata/repository/domain/syncState/SyncStateEntity;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun select (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
 public final class io/getstream/chat/android/livedata/repository/domain/user/UserDao_Impl : io/getstream/chat/android/livedata/repository/domain/user/UserDao {
