+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/domain/syncState/SyncStateDao.kt
@@ -1,10 +1,9 @@
-package io.getstream.chat.android.livedata.dao
+package io.getstream.chat.android.livedata.repository.domain.syncState
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import io.getstream.chat.android.livedata.entity.SyncStateEntity
 
 @Dao
 internal interface SyncStateDao {
