+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/repository/helper/BaseRepositoryFacadeTest.kt
@@ -6,9 +6,9 @@ import io.getstream.chat.android.livedata.repository.ChannelConfigRepository
 import io.getstream.chat.android.livedata.repository.ChannelRepository
 import io.getstream.chat.android.livedata.repository.QueryChannelsRepository
 import io.getstream.chat.android.livedata.repository.RepositoryFacade
-import io.getstream.chat.android.livedata.repository.SyncStateRepository
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepository
 import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionRepository
+import io.getstream.chat.android.livedata.repository.domain.syncState.SyncStateRepository
 import io.getstream.chat.android.livedata.repository.domain.user.UserRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineScope
