+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/utils/TestDataHelper.kt
@@ -39,7 +39,7 @@ import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.livedata.controller.QueryChannelsSpec
-import io.getstream.chat.android.livedata.entity.SyncStateEntity
+import io.getstream.chat.android.livedata.repository.domain.syncState.SyncStateEntity
 import io.github.cdimascio.dotenv.dotenv
 import java.time.Instant
 import java.util.Calendar
